+++ src/Helper/DenyAccessUnlessCheckSignature.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\BlobBundle\Helper;
 
+use Dbp\Relay\BlobLibrary\Helpers\SignatureTools;
 use Dbp\Relay\CoreBundle\Exception\ApiError;
 use Jose\Component\Core\AlgorithmManager;
 use Jose\Component\Core\JWK;
@@ -28,9 +29,7 @@ class DenyAccessUnlessCheckSignature
      */
     public static function create(string $secret, array $payload): string
     {
-        $jwk = self::createJWK($secret);
-
-        return self::generateToken($jwk, $payload);
+        return SignatureTools::create($secret, $payload);
     }
 
     /**
@@ -46,10 +45,10 @@ class DenyAccessUnlessCheckSignature
      */
     public static function verify(string $secret, string $token): array
     {
-        $jwk = self::createJWK($secret);
+        $jwk = SignatureTools::createJWK($secret);
         $payload = [];
 
-        if (!self::verifyToken($jwk, $token, $payload)) {
+        if (!SignatureTools::verifyToken($jwk, $token, $payload)) {
             /* @noinspection ForgottenDebugOutputInspection */
             //dump(['token' => $token, 'payload' => $payload, 'secret' => $secret]);
             throw ApiError::withDetails(Response::HTTP_FORBIDDEN, 'Signature invalid', 'blob:signature-invalid');
@@ -100,70 +99,6 @@ class DenyAccessUnlessCheckSignature
         $url = explode('&sig=', $request->getRequestUri());
 
         // generate hmac sha256 hash over the uri except the signature part
-        return hash('sha256', $url[0]);
-    }
-
-    /**
-     * Create the JWK from a shared secret.
-     *
-     * @param string $secret to create the (symmetric) JWK from
-     */
-    public static function createJWK(string $secret): JWK
-    {
-        return JWKFactory::createFromSecret(
-            $secret,
-            [
-                'alg' => 'HS256',
-                'use' => 'sig',
-            ]
-        );
-    }
-
-    /**
-     * Generate the token.
-     *
-     * @param JWK   $jwk     json web key
-     * @param array $payload as json string to secure
-     *
-     * @return string secure token
-     *
-     * @throws \JsonException
-     */
-    public static function generateToken(JWK $jwk, array $payload): string
-    {
-        $algorithmManager = new AlgorithmManager([new HS256()]);
-        $jwsBuilder = new JWSBuilder($algorithmManager);
-
-        $jws = $jwsBuilder
-            ->create()
-            ->withPayload(json_encode($payload, JSON_THROW_ON_ERROR))
-            ->addSignature($jwk, ['alg' => 'HS256'])
-            ->build();
-
-        return (new CompactSerializer())->serialize($jws, 0);
-    }
-
-    /**
-     * Verify a JWS token.
-     *
-     * @param string $token   the JWS token as string
-     * @param array  $payload to extract from token on success
-     *
-     * @throws \JsonException
-     */
-    public static function verifyToken(JWK $jwk, string $token, array &$payload): bool
-    {
-        $algorithmManager = new AlgorithmManager([new HS256()]);
-        $jwsVerifier = new JWSVerifier($algorithmManager);
-        $serializerManager = new JWSSerializerManager([new CompactSerializer()]);
-        $jws = $serializerManager->unserialize($token);
-
-        if ($ok = $jwsVerifier->verifyWithKey($jws, $jwk, 0)) {
-            $payload = json_decode($jws->getPayload(), true, 512, JSON_THROW_ON_ERROR);
-        }
-//        $ok = $jwsVerifier->verifyWithKey($jws, $jwk, 0);
-//        $payload = json_decode($jws->getPayload(), true, 512, JSON_THROW_ON_ERROR);
-
-        return $ok;
+        return SignatureTools::generateSha256Checksum($url[0]);
     }
 }