+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.1.10
+* Use `dbp/relay-blob-library` and move some code from `\Dbp\Relay\BlobBundle\Helper\DenyAccessUnlessCheckSignature`
+  to `\Dbp\Relay\BlobLibrary\Helpers\SignatureTools`
+
 # v0.1.9
 * Change meaning of `binary` GET parameter
   * before it meant that a 302 redirect should be the answer